Figured out what reason was
The reason was because shuffle inserts a element into the document that is 10 px wide. It then uses JS to calc if this is 10px. Since I am using a flex body with 100% width, this messes up that calculation, which makes the calculated column width of a column more than it should be. This is prob Shuffle's fault, but it was an interesting problem to debug, since I had a codepen that was almost identical to this project that worked.

To fix the problem, I just added some code in the shuffle js to add the element to the main block element instead of as a sibling of that.